+++ tests/conftest.py
@@ -27,6 +27,9 @@ logger = logging.getLogger(__name__)
 class Sender:
     """Sender class"""
 
+    def __str__(self):
+        return f"Sender(id={id(self)})"
+
     @signal
     def value_changed(self, value):
         """Signal for value changes"""
@@ -39,6 +42,9 @@ class Sender:
 @with_signals
 class Receiver:
     """Receiver class"""
+
+    def __str__(self):
+        return f"Receiver(id={self.id})"
 
     def __init__(self):
         super().__init__()
@@ -105,10 +111,8 @@ def setup_logging():
 
     # Setting to WARNING level by default
     default_level = logging.WARNING
-    # default_level = logging.DEBUG
 
     # Can enable DEBUG mode via environment variable
-
     if os.environ.get("PYNNEX_DEBUG"):
         default_level = logging.DEBUG
 
@@ -125,7 +129,8 @@ def setup_logging():
 
     # Setting formatter
     formatter = logging.Formatter(
-        "%(asctime)s - %(name)s - %(levelname)s - %(message)s", datefmt="%H:%M:%S"
+        "%(asctime)s - %(name)s - [%(filename)s:%(funcName)s] - %(levelname)s - %(message)s",
+        datefmt="%H:%M:%S"
     )
     console_handler.setFormatter(formatter)
     root.addHandler(console_handler)
@@ -133,3 +138,19 @@ def setup_logging():
     # Setting package logger levels
     logging.getLogger("pynnex").setLevel(default_level)
     logging.getLogger("tests").setLevel(default_level)
+
+    # Disable trace logging
+    logging.getLogger("pynnex.signal").setLevel(logging.WARNING)
+    logging.getLogger("pynnex.slot").setLevel(logging.WARNING)
+    logging.getLogger("pynnex.signal.trace").setLevel(logging.WARNING)
+    logging.getLogger("pynnex.slot.trace").setLevel(logging.WARNING)
+
+    # For debugging
+    root.setLevel(logging.DEBUG)
+    console_handler.setLevel(logging.DEBUG)
+    logging.getLogger("pynnex").setLevel(logging.DEBUG)
+    logging.getLogger("tests").setLevel(logging.DEBUG)
+    logging.getLogger("pynnex.signal").setLevel(logging.DEBUG)
+    logging.getLogger("pynnex.slot").setLevel(logging.DEBUG)
+    logging.getLogger("pynnex.signal.trace").setLevel(logging.DEBUG)
+    logging.getLogger("pynnex.slot.trace").setLevel(logging.DEBUG)